refactor: rearrange modal files
