+++ src/components/modals/EditTask.js
@@ -14,11 +14,11 @@ const customStyles = {
 
 Modal.setAppElement("*");
 
-export default function EditModal({
+export default function EditTask({
   modalStatus,
   setModalStatus,
-  projectName,
-  setProjectName,
+  taskName,
+  setTaskName,
   handleUpdate,
 }) {
   let subtitle;
@@ -32,8 +32,10 @@ export default function EditModal({
   }
 
   const handleEditProject = () => {
-    handleUpdate();
-    closeModal();
+    if (taskName) {
+      handleUpdate();
+      closeModal();
+    }
   };
 
   return (
@@ -50,15 +52,15 @@ export default function EditModal({
             className="capitalize"
             ref={(_subtitle) => (subtitle = _subtitle)}
           >
-            Edit Project Title
+            Edit Task
           </h2>
           <button onClick={closeModal}>X</button>
         </div>
         <div>
           <input
             type="text"
-            value={projectName}
-            onChange={({ target }) => setProjectName(target.value)}
+            value={taskName}
+            onChange={({ target }) => setTaskName(target.value)}
             className="w-full px-2 py-1.5 rounded my-2 border border-gray-primary bg-white focus:outline-none"
             placeholder="Name your project"
           />
